Implement fetching a task by id
Our goal here is to make the server-side ID match the client side's
local storage index. Start with 0, new items that get "pushed" onto the
list will get the next number, and so on.

This isn't what I'd choose for a serious app, but it will be simple
enough for this example app.